requirements.txt
Flask
pymongo
python-chess